requirements.txt
# Pinned for reproducibility
pytest==8.4.1
pytest-timeout==2.3.1
typing-extensions==4.15.0
fastapi==0.116.1
pydantic==2.11.7
httpx==0.28.1
uvicorn==0.35.0
gunicorn==22.0.0
click==8.2.1
fpdf2==2.8.4
prometheus-client==0.20.0
